+++ posthog-react-native/src/posthog-rn.ts
@@ -1,6 +1,7 @@
 import { AppState, Dimensions, Linking } from 'react-native'
 
 import {
+  PostHogCaptureOptions,
   PostHogCore,
   PosthogCoreOptions,
   PostHogFetchOptions,
@@ -170,16 +171,20 @@ export class PostHog extends PostHogCore {
   }
 
   // Custom methods
-  screen(name: string, properties?: any): this {
+  screen(name: string, properties?: { [key: string]: any }, options?: PostHogCaptureOptions): this {
     // Screen name is good to know for all other subsequent events
     this.registerForSession({
       $screen_name: name,
     })
 
-    return this.capture('$screen', {
-      ...properties,
-      $screen_name: name,
-    })
+    return this.capture(
+      '$screen',
+      {
+        ...properties,
+        $screen_name: name,
+      },
+      options
+    )
   }
 
   initReactNativeNavigation(options: PostHogAutocaptureOptions): boolean {